requirements.txt
aiogram==3.5.0
httpx==0.27.0
google-generativeai>=0.7.0,<1.0.0
anthropic==0.25.0
python-dotenv==1.0.0
loguru==0.7.2
pydantic==2.5.0
aiofiles==23.2.1
pyyaml==6.0.1